pyproject.toml
[project]
name = "eagent"
version = "0.1.1"
description = "Production-ready LangGraph agent example"
readme = "README.md"
authors = [
    { name = "liuyan", email = "lyfx99@icloud.com" }
]
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core",
    "langgraph",
    "langsmith",
    "langchain-docling>=0.2.0",
    "pydantic>=2",
    "pydantic-settings>=2",
    "typer",
    "rich",
    "python-dotenv",
    "langchain-anthropic>=1.2.0",
    "langchain-openai>=1.1.0",
    "langchain-community>=0.4.1",
    "docling>=2.63.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "faiss-cpu>=1.13.1",
    "pkuseg>=0.0.25",
    "jieba>=0.42.1",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
]
visual = [
    "gradio>=4.0.0",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
]

[project.scripts]
rob2 = "cli.app:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
[tool.uv.extra-build-dependencies]
      pkuseg = ["numpy"]
